fix violations
git-svn-id: http://opensource.adobe.com/svn/opensource/flexpmd/trunk@13178 a9308255-753e-0410-a2e9-80b3fbc4fff6

+++ flex-pmd-metrics/src/main/java/com/adobe/ac/pmd/metrics/AverageClassMetrics.java
@@ -41,7 +41,7 @@ public class AverageClassMetrics extends AverageMetricsBase
    {
       super( totalClassNumber, asDocs, nonCommentStatements );
       averageFunctions = totalClassNumber > 0 ? functions
-                                                   / totalClassNumber
+                                                   / ( double ) totalClassNumber
                                              : 0;
    }
